requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
greenlet==0.4.15
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mongoengine==0.20.0
mutagen==1.42.0
PyJWT==1.7.1
pymongo==3.10.1
requests==2.23.0
six==1.12.0
urllib3==1.25.9
Werkzeug==0.16.0
